Fix for unregistering service when re-advertising
When we change the properties that we’re advertising over mDNS we unregister and re-register the service. However, we were calling unregister after changing the properties. This happened to work with our version of jmdns, which only compared the qualified name. jmdns has quite a few forks however, and it’s best to adhere to the implied contract when calling unregisterService.

This is adapted from a fix proposed by J-N-K