keep subscriptions up to date when accessories are replaced
but the connection is _not_ torn down

also add the ability to modify accessories in a batch without
having to regenerate all AIDs and IIDs (and other temporary
data structures) at each intermediate step. You must use this
mechanism if you plan on replacing an existing accessory with an
equivalent one, but want to keep subscriptions intact.