refactor pairing to match HAP spec more closely
mostly renames, but significant change is that when we don't recognize
a user, return an error code instead of throwing an exception and returning
a 500 with no content.

several responses are now generated with PairingResponse, rather than as
a raw byte array, so that the reader can more easily follow what the response
is.

several more trace loggings showing the details of the pairing process have
been added, for debug use

also ensure that we check if the MAC has changed when refreshing auth info